pyproject.toml
[project]
name = "spam-detection"
version = "0.1.0"
description = "IMAP-integrated ML spam detection service with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "imapclient>=3.0.1",
    "cryptography>=41.0.7",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "celery[redis]>=5.3.4",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "joblib>=1.3.2",
    "nltk>=3.8.1",
    "scikit-learn==1.6.1",
    "sse-starlette>=1.8.2",
    "aiosmtplib>=3.0.1",
    "email-validator>=2.1.0",
    "python-multipart>=0.0.6",
    "alembic>=1.13.0",
    "redis>=5.0.1",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
]
